% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_sum.R
\name{cyl_sum}
\alias{cyl_sum}
\title{Make a specific summary table (demo)}
\usage{
cyl_sum(x)
}
\arguments{
\item{x}{The var in mtcars that we want average of}
}
\value{
A df summarizing mtcars by cyl
}
\description{
Make a specific summary table (demo)
}
