% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{hexbin}
\alias{hexbin}
\title{Create hex plot}
\usage{
hexbin(df, x, y, bins = 10)
}
\arguments{
\item{df}{A df}

\item{x}{Name of col}

\item{y}{Name of col}

\item{bins}{Numeric value}
}
\value{
A plot
}
\description{
Demo
}
